backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiobotocore==2.25.1
    # via s3fs
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   -r requirements.in
    #   aiobotocore
    #   s3fs
aioitertools==0.12.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
botocore==1.40.61
    # via aiobotocore
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   netcdf4
    #   requests
cftime==1.6.5
    # via
    #   -r requirements.in
    #   netcdf4
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   dask
    #   uvicorn
cloudpickle==3.1.1
    # via dask
contourpy==1.3.3
    # via matplotlib
crc32c==2.8
    # via numcodecs
cycler==0.12.1
    # via matplotlib
dask==2025.10.0
    # via -r requirements.in
donfig==0.8.1.post1
    # via zarr
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.120.4
    # via -r requirements.in
fonttools==4.60.1
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via
    #   -r requirements.in
    #   dask
    #   kerchunk
    #   s3fs
geographiclib==2.0
    # via
    #   -r requirements.in
    #   geopy
geopy==2.4.1
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h5netcdf==1.7.3
    # via -r requirements.in
h5py==3.15.1
    # via h5netcdf
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jmespath==1.0.1
    # via
    #   aiobotocore
    #   botocore
kerchunk==0.2.9
    # via -r requirements.in
kiwisolver==1.4.9
    # via matplotlib
locket==1.0.0
    # via partd
loguru==0.7.3
    # via -r requirements.in
matplotlib==3.10.7
    # via -r requirements.in
multidict==6.7.0
    # via
    #   aiobotocore
    #   aiohttp
    #   yarl
netcdf4==1.7.3
    # via -r requirements.in
numcodecs[crc32c]==0.16.3
    # via
    #   kerchunk
    #   zarr
numpy==2.3.4
    # via
    #   -r requirements.in
    #   cftime
    #   contourpy
    #   h5py
    #   kerchunk
    #   matplotlib
    #   netcdf4
    #   numcodecs
    #   pandas
    #   patsy
    #   scipy
    #   statsmodels
    #   xarray
    #   zarr
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   dask
    #   h5netcdf
    #   matplotlib
    #   statsmodels
    #   xarray
    #   zarr
pandas==2.3.3
    # via
    #   -r requirements.in
    #   statsmodels
    #   xarray
partd==1.4.2
    # via dask
patsy==1.0.2
    # via statsmodels
pyarrow==20.0.0
    # via -r requirements.in
pillow==12.0.0
    # via
    #   -r requirements.in
    #   matplotlib
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psutil==7.1.3
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.4
    # via pydantic
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   aiobotocore
    #   botocore
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   dask
    #   donfig
    #   uvicorn
requests==2.32.5
    # via -r requirements.in
s3fs==2025.10.0
    # via -r requirements.in
scipy==1.16.3
    # via
    #   -r requirements.in
    #   statsmodels
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via -r requirements.in
starlette==0.49.3
    # via fastapi
statsmodels==0.14.6
    # via -r requirements.in
toolz==1.1.0
    # via
    #   dask
    #   partd
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   numcodecs
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
    #   zarr
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
ujson==5.11.0
    # via
    #   -r requirements.in
    #   kerchunk
urllib3==2.5.0
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.38.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.3
    # via aiobotocore
xarray==2025.10.1
    # via -r requirements.in
xlsxwriter==3.2.9
    # via -r requirements.in
yarl==1.22.0
    # via aiohttp
zarr==3.1.3
    # via
    #   -r requirements.in
    #   kerchunk
